+++ cvat-ui/src/reducers/root-reducer.ts
@@ -19,6 +19,7 @@ import userAgreementsReducer from './useragreements-reducer';
 import reviewReducer from './review-reducer';
 import exportReducer from './export-reducer';
 import cloudStoragesReducer from './cloud-storages-reducer';
+import organizationsReducer from './organizations-reducer';
 
 export default function createRootReducer(): Reducer {
     return combineReducers({
@@ -38,5 +39,6 @@ export default function createRootReducer(): Reducer {
         review: reviewReducer,
         export: exportReducer,
         cloudStorages: cloudStoragesReducer,
+        organizations: organizationsReducer,
     });
 }