+++ cvat/apps/dataset_repo/views.py
@@ -5,7 +5,7 @@
 from django.http import HttpResponseBadRequest, JsonResponse
 from rules.contrib.views import permission_required, objectgetter
 
-from cvat.apps.authentication.decorators import login_required
+from cvat.apps.iam.decorators import login_required
 from cvat.apps.engine.log import slogger
 from cvat.apps.engine import models
 from cvat.apps.dataset_repo.models import GitData
@@ -55,8 +55,6 @@ def create(request, tid):
 
 
 @login_required
-@permission_required(perm=['engine.task.access'],
-    fn=objectgetter(models.Task, 'tid'), raise_exception=True)
 def push_repository(request, tid):
     try:
         slogger.task[tid].info("push repository request")
@@ -75,8 +73,6 @@ def push_repository(request, tid):
 
 
 @login_required
-@permission_required(perm=['engine.task.access'],
-    fn=objectgetter(models.Task, 'tid'), raise_exception=True)
 def get_repository(request, tid):
     try:
         slogger.task[tid].info("get repository request")