+++ cvat/apps/lambda_manager/apps.py
@@ -2,4 +2,4 @@ from django.apps import AppConfig
 
 
 class LambdaManagerConfig(AppConfig):
-    name = 'lambda_manager'
+    name = 'cvat.apps.lambda_manager'