+++ tests/cypress/integration/actions_tasks/issue_2473_import_annotations_frames_dots_in_name.js
@@ -105,7 +105,7 @@ context('Import annotations for frames with dots in name.', { browser: '!firefox
                         .get('input[type=file]')
                         .attachFile(annotationArchiveName);
                 });
-            cy.intercept('GET', '/api/v1/jobs/**/annotations').as('uploadAnnotationsGet');
+            cy.intercept('GET', '/api/v1/jobs/**/annotations?**').as('uploadAnnotationsGet');
             confirmUpdate('.cvat-modal-content-load-job-annotation');
             cy.wait('@uploadAnnotationsGet').its('response.statusCode').should('equal', 200);
             cy.get('.cvat-notification-notice-upload-annotations-fail').should('not.exist');