+++ tests/rest_api/config.py
@@ -1,0 +1,9 @@
+# Copyright (C) 2021 Intel Corporation
+#
+# SPDX-License-Identifier: MIT
+
+import os
+
+ROOT_DIR = os.path.dirname(__file__)
+ASSETS_DIR = os.path.join(ROOT_DIR, 'assets')
+USER_PASS = '!Q@W#E$R'